abstract oauth adapters
