+++ src/Appwrite/Auth/OAuth2.php
@@ -50,13 +50,6 @@ abstract class OAuth2
             $this->addScope($scope);
         }
     }
-    
-    /**
-     * Check if the OAuth credentials are valid
-     *
-     * @throws \Exception
-     */
-    abstract public function verifyCredentials(): void;
 
     /**
      * @return string
