+++ src/Appwrite/Auth/OAuth2/Discord.php
@@ -184,8 +184,4 @@ class Discord extends OAuth2
 
         return $this->user;
     }
-
-    public function verifyCredentials(): void {
-        // TODO: Implement, eventuelly. Refer to GitHub.php in this directory for inspiration
-    }
 }
